@@ -4,40 +4,39 @@
     "type": "project",
     "license": "MIT",
     "require": {
-        "php": "^7.4",
-        "ext-gd": "*",
-        "ext-json": "*",
-        "ext-pdo": "*",
-        "bensampo/laravel-enum": "^v3.0.0",
         "beyondcode/laravel-websockets": "^1.4",
         "cebe/php-openapi": "^1.3",
         "creativeorange/gravatar": "~1.0",
         "cviebrock/laravel-elasticsearch": "^8.0.0",
         "darkaonline/l5-swagger": "^8.0.0",
         "doctrine/dbal": "^2.10",
+        "douglasroos/laravel-settings": "1.0.11",
+        "ext-gd": "*",
+        "ext-json": "*",
+        "ext-pdo": "*",
         "fideloper/proxy": "^4.0",
         "hisorange/browser-detect": "^4.0",
         "laravel/framework": "^8.0",
         "laravel/sanctum": "^2.0",
-        "laravel/socialite": "^v5.0.0",
+        "laravel/socialite": "^5",
         "laravel/tinker": "^2.0",
         "laravel/ui": "^3.0",
         "marvinlabs/laravel-discord-logger": "^1.1",
         "monicahq/laravel-cloudflare": "^1.6",
         "nojacko/email-validator": "^1.1",
         "ongr/elasticsearch-dsl": "^7.2",
-        "qcod/laravel-settings": "^1.0",
-        "restcord/restcord": "^0.4.0",
+        "php": "~8.1",
+        "restcord/restcord": "^0.5.0",
         "seinopsys/rgbacolor": "^1.0",
-        "socialiteproviders/deviantart": "^4.0",
-        "socialiteproviders/discord": "^2.0",
+        "socialiteproviders/deviantart": "^4.1.0",
+        "socialiteproviders/discord": "^4.1.1",
         "spatie/eloquent-sortable": "^3.8",
         "spatie/laravel-activitylog": "^3.14",
         "spatie/laravel-medialibrary": "^8.0.0",
         "spatie/laravel-responsecache": "^6.6",
         "stechstudio/laravel-php-cs-fixer": "^2.0",
         "valorin/pwned-validator": "^1.3",
-        "webpatser/laravel-uuid": "^3.0"
+        "webpatser/laravel-uuid": "^4.0.1"
     },
     "require-dev": {
         "barryvdh/laravel-ide-helper": "^2.6",
@@ -46,7 +45,7 @@
         "mockery/mockery": "^1.0",
         "nunomaduro/collision": "^5.0",
         "phpunit/phpunit": "^9.0",
-        "roave/security-advisories": "dev-master"
+        "roave/security-advisories": "dev-latest"
     },
     "config": {
         "optimize-autoloader": true,
@@ -58,6 +57,12 @@
             "dont-discover": []
         }
     },
+    "repositories": [
+        {
+            "type": "github",
+            "url": "https://github.com/douglasroos/laravel-settings"
+        }
+    ],
     "autoload": {
         "psr-4": {
             "App\\": "app/",
